+++ lib/danger/ci_source/circle.rb
@@ -30,6 +30,9 @@ module Danger
     def self.validates_as_pr?(env)
       return true if env["CI_PULL_REQUEST"]
 
+      # Real-world talk, it should be worrying if none of these are in the environment
+      return false unless ["CIRCLE_CI_API_TOKEN", "CIRCLE_PROJECT_USERNAME", "CIRCLE_PROJECT_REPONAME", "CIRCLE_BUILD_NUM"].all? { |x| env[x] }
+
       # Uses the Circle API to determine if it's a PR otherwose
       @circle_token = env["CIRCLE_CI_API_TOKEN"]
       !pull_request_url(env).nil?
@@ -62,14 +65,17 @@ module Danger
     def initialize(env)
       self.repo_url = GitRepo.new.origins # CircleCI doesn't provide a repo url env variable :/
 
-      @circle_token = env["CIRCLE_CI_API_TOKEN"]
-      url = pull_request_url(env)
+      pr_url = env["CI_PULL_REQUEST"]
+      unless pr_url
+        @circle_token = env["CIRCLE_CI_API_TOKEN"]
+        pr_url = pull_request_url(env)
+      end
 
-      if URI.parse(url).path.split("/").count == 5
-        paths = URI.parse(url).path.split("/")
+      pr_path = URI.parse(pr_url).path.split("/")
+      if pr_path.count == 5
         # The first one is an extra slash, ignore it
-        self.repo_slug = paths[1] + "/" + paths[2]
-        self.pull_request_id = paths[4]
+        self.repo_slug = pr_path[1] + "/" + pr_path[2]
+        self.pull_request_id = pr_path[4]
       end
     end
   end